+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # KAMI: Reborn Changelog
+## 1.0.4
+### Added
+- Added a config option to make the Ichorium material overpowered (Tinkers' Construct / Construct's Armory).
+- Added a config option to make the Ichorium tool material use a more overpowered trait (Tinkers' Construct).
+
+---
+
 ## 1.0.3
 ### Fixed
 - Fixed a ton of server crashes.
@@ -7,7 +14,7 @@
 
 ## 1.0.2
 ### Fixed
-- Fixed a potential client crash with the Tinkers' Construct integration
+- Fixed a potential client crash with the Tinkers' Construct integration.
 
 ---